Fixed the problem where bullet instantly disappear on first shot
When each pooled bullets became active for the first time, OnEnable was getting called before Start and because of m_StayTime not being set and initiated as 0, it invoked Remove function after the time of 0f - which was instant.
Variable initiation routine is moved to Awake as it is called before OnEnable.